requirements.txt
boto3==1.34.34
findspark==2.0.1
pandas==2.0.0
numpy==1.24.2
s3fs==2024.3.1
fsspec==2024.3.1
kafka-python==2.0.2
lz4==4.3.3
pydantic==2.8.2
pydantic-settings==2.3.4
pyarrow==17.0.0